Forked Path Movement
Game waits for player to press arrow key to select path when at a fork and the player still has spaces left to move. Reworked how player's choice is registered, which fixed desync bug from previous commit.

I changed the debug textboxes to display when the game is waiting for user input.