Working GameOver Behavior & Debugging QOL
Game properly continues until 2 players reach the end; fixed issue where TargetRpc was being sent to wrong player.  Added spashscreens for when the game is getting ready and when the game is over; sync issue but these are placeholder screens anyways.

Added field on player prefab to allow for debug mode; this lets you tweak aspects of the game while being able to return to playing the game normally without recompiling scripts.